Counting pairs
Two pairs of integers (a, b) and (c, d) are considered distinct if
at least one element of (a, b) ∉ (c, d). For example given a list
(1, 2, 2, 3), (1, 2) is distinct from (1, 3) and (2, 3) but not from
(1, 2) with 2 chosen from a different index in the list. A pair is
valid if a ≤ b.
You will be given an integer k and a list of integers. Count the
number of distinct valid pairs of integers (a, b) in the list for
which a + k = b.